requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pytz==2021.1
requests==2.25.1
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.11
SQLAlchemy-serializer==1.3.4.4
urllib3==1.26.4
vk-api==11.9.1
Werkzeug==1.0.1
WTForms==2.3.3
